engine: clean up rocks mod
